First pass Plugin.
Presently it seems to commit to the history correctly (provided that
g:vlh_repository_dir is set).

No mechanism exists for viewing the history, showing diffs, etc.

Some cleanup should be done before installing the plugin.  At present, a
whole bunch of files would be dumped into ~/.vim/plugin.  It would be
better for all the ancillary files to be in a subdirectory.