Fixed sudo issue, VLHOpen/Diff ft issue.
Saved files are now chmod'd to 0600 and chown'd to the user that owns the
repository location (presumably the user).  This means that a save of
sudo vim /etc/hosts will result in a backup file not owned by root, and so
won't cause problems during, e.g. `git gc`.

VLHOpen and VLHDiff now set the filetype of the temp file to match the
filetype of the current buffer, which gives syntax highlighting, etc.

VLHDiff now uses wincmd x to switch the windows, instead of rotating.
This works much better when you have more than one window open when
running VLHDiff (e.g. if you have project or tagslist open).

Reorganized some priorities in README.